# apps/ssi-abstraction/swagger.json
{
  "components": {
    "schemas": {
...
      "SchemaTemplate": {
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "pattern": "^(\\d+\\.)?(\\d+\\.)?(\\*|\\d+)$",
            "type": "string"
          },
          "attributes": {
            "items": {},
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "name",
          "version",
          "attributes"
        ]
      },
...
    }
  },
  "info": {
    "title": "ssi-abstraction-agent",
    "version": "0.8.0",
    "description": "Rest endpoint wrapper for using your agent over HTTP"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/schemas/{schemaId}": {
      "get": {
        "operationId": "SchemaController.getSchemaById",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get schema by id",
        "tags": [
          "Schema"
        ]
      }
    },
    "/schemas/": {
      "post": {
        "operationId": "SchemaController.createSchema",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaTemplate"
              }
            }
          },
          "description": "SchemaTemplate",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Create schema",
        "tags": [
          "Schema"
        ]
      }
    },
...
  }
}